refactor: move dataService delete to revisionService
