+++ EMS/core-bundle/config/controllers.xml
@@ -151,6 +151,7 @@
         <service id="EMS\CoreBundle\Controller\ContentManagement\DataController" public="true">
             <argument type="service" id="logger" />
             <argument type="service" id="ems.service.data" />
+            <argument type="service" id="ems.service.revision" />
             <argument type="service" id="ems.service.search" />
             <argument type="service" id="ems.service.contenttype" />
             <argument type="service" id="ems.service.environment" />
