+++ EMS/core-bundle/config/services.xml
@@ -234,6 +234,7 @@
             <argument type="service" id="ems_core.core_mail.mailer_service" />
             <argument type="service" id="ems.repository.user" />
             <argument type="service" id="security.user_password_hasher" />
+            <argument type="service" id="security.authorization_checker" />
             <argument>%ems_core.template_namespace%</argument>
         </service>
 
@@ -283,6 +284,7 @@
             <argument type="service" id="form.factory" />
             <argument type="service" id="logger" />
             <argument type="service" id="emsco.logger.audit" />
+            <argument type="service" id="Doctrine\ORM\EntityManagerInterface" />
             <argument type="service" id="EMS\CoreBundle\Repository\RevisionRepository" />
             <argument type="service" id="ems.service.publish" />
             <argument type="service" id="EMS\CoreBundle\Service\ContentTypeService"/>
