+++ EMS/core-bundle/src/Core/Component/MediaLibrary/MediaLibraryService.php
@@ -85,7 +85,7 @@ class MediaLibraryService
     public function deleteDocument(MediaLibraryDocument $mediaDocument, ?string $username = null): void
     {
         $document = $mediaDocument->document;
-        $this->dataService->delete($document->getContentType(), $document->getOuuid(), $username);
+        $this->revisionService->delete($document->getContentType(), $document->getOuuid(), $username);
     }
 
     public function exists(MediaLibraryFile|MediaLibraryFolder $document): bool
